requirements.txt
# CyAnno Pipeline Requirements
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
snakemake>=6.0.0
pyyaml>=5.4.0
